+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Upcoming Changes (unreleased)
 - Added method ```valid_country``` for returning country for parsed phone just in case phone number was valid
 - Changed behavior of method ```country```, now it returns main country for international code in case there is such country, or first country from array
+- Added ```local_number``` method to ```Phone```, returns local number without area code
+- Added ```area_code``` method to ```Phone```, returns area code of phone or nil if none present for parsed number
 
 ## 0.4.9 - July 28, 2015